Documentation updated Cache CacheExpirationTimeMinutes changed to double to handle minutes and seconds

+++ Domain/Entities/AppSettings.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities
     public class AppSettings
     {
         public required string BestStoryIdsCacheName { get; set; }
-        public required int CacheExpirationTimeMinutes { get; set; }
+        public required double CacheExpirationTimeMinutes { get; set; }
         public required string DateFormat { get; set; }
     }
 }
